# apps/server/openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Managed Agents Workbench API",
		"version": "1.0.0"
	},
	"components": {
		"schemas": {
...
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"error": {
						"type": "object",
						"properties": {
							"message": {
								"type": "string"
							}
						},
						"required": ["message"]
					}
				},
				"required": ["error"]
			},
...
			"ProjectVersioningStatus": {
				"type": "object",
				"properties": {
					"initialized": {
						"type": "boolean"
					},
					"enabled": {
						"type": "boolean"
					},
					"store_root": {
						"type": "string"
					},
					"config_path": {
						"type": "string"
					},
					"head_version": {
						"type": "string",
						"nullable": true
					},
					"source_status": {
						"type": "string",
						"enum": ["clean", "modified", "unversioned"]
					},
					"source_versioned": {
						"type": "boolean"
					},
					"write_blockers": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"restore_blockers": {
						"type": "array",
						"items": {
							"type": "string"
						}
					}
				},
				"required": [
					"initialized",
					"enabled",
					"store_root",
					"config_path",
					"head_version",
					"source_status",
					"source_versioned",
					"write_blockers",
					"restore_blockers"
				]
			},
			"ProjectVersionsResponse": {
				"type": "object",
				"properties": {
					"versions": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProjectVersion"
						}
					},
					"next_cursor": {
						"type": "string",
						"nullable": true
					}
				},
				"required": ["versions", "next_cursor"]
			},
			"ProjectVersion": {
				"type": "object",
				"properties": {
					"version_id": {
						"type": "string"
					},
					"short_version": {
						"type": "string"
					},
					"parent_version": {
						"type": "string",
						"nullable": true
					},
					"source_hash": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"created_by": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					}
				},
				"required": [
					"version_id",
					"short_version",
					"parent_version",
					"source_hash",
					"message",
					"created_by",
					"created_at"
				]
			},
...
		},
...
	},
	"paths": {
...
		"/api/project/versioning": {
			"get": {
				"responses": {
					"200": {
						"description": "Local directory source snapshot store and versioning status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectVersioningStatus"
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable entity",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/api/project/versions": {
			"get": {
				"parameters": [
					{
						"schema": {
							"type": "string"
						},
						"required": false,
						"name": "cursor",
						"in": "query"
					},
					{
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 100
						},
						"required": false,
						"name": "limit",
						"in": "query"
					}
				],
				"responses": {
					"200": {
						"description": "Local directory source snapshots",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectVersionsResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable entity",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	}
}